Implementing Comparator classes for Sortare button
Clicking the button Sortare opens a new window that allows the user to 
choose the column on which to sort the table in an ascending order.

Clicking on a table column header will also sort the table on the first
click and descending on the second click in the same column header.